+++ platform/webex/option_schema.go
@@ -1,0 +1,7 @@
+package webex
+
+import "github.com/timmyagentic/cc-connect-next/core"
+
+func init() {
+	core.RegisterPlatformConfigOptions("webex", core.DescribePlatformOptions([]string{"allow_from", "token"}))
+}